Refactored files into module
